refactor: Update MineralListItem type to include MineralListItem[] in associates property
The MineralListItem type in the `types/types.ts` file has been updated to include an array of MineralListItem objects in the associates property. This change allows for more flexibility in representing associated minerals. By including the option to have multiple associates, we improve the accuracy and completeness of the data model.