refactor: Update assignedBy field in PhotoOnMineral model to be optional
This commit modifies the `assignedBy` field in the `PhotoOnMineral` model of the `prisma/schema.prisma` file to be optional by adding a question mark after the field type. This change allows for cases where the `assignedBy` field may not have a value. The purpose of this update is to provide flexibility in the data model and accommodate scenarios where the `assignedBy` field is not always required.