refactor: Update Home page CTA based on user session
Update the call-to-action button on the Home page to dynamically display different text based on whether the user is logged in or not. If the user is logged in, the button will now say "Go to Settings". If the user is not logged in, the button will say "Log In or Sign Up". This change improves the user experience by providing clearer and more relevant options for different user states.